Phone: Ignore touch during network operators search
When network search is in progress, the progress dialog is
dismissed if user taps the screen. This in-turn cancels
manual search and switches to the previous screen.
This behavior can be avoided if the dialog property is set
to ignore user touch events.
If required BACK key can be used to cancel search and move
to previous screen.

CRs-Fixed: 367631
(cherry picked from commit 1f1d395015fd8b662ee0e40876e44f9e1d6ceb57)

Change-Id: Ib53de0e4c9a379474ccd1889afa4b4dba7be1a7d